Small fixes to existing generators
